Split server run command tests by OS
